[refactor] : SortingModalBottomView CTAButton 새로운 생성자 적용
